requirements.txt
atlassian-python-api==3.41.4
python-dotenv==1.0.0
gigachat==0.1.39.post1
numpy==2.2.4
faiss-cpu==1.10.0


mcp~=1.5.0
requests~=2.32.3